git-svn-id: http://www.mz3.jp/svn/mz3/trunk@1295 875e8a4c-fb2d-0410-984e-92a9dd373959

+++ src/MZ3/MZ3View.cpp
@@ -6770,7 +6770,7 @@ void CMZ3View::OnCategorymenuMoveDown()
 /// ƒJƒeƒSƒŠ‚ÌˆÚ“®
 void CMZ3View::MoveCategoryItem(int oldCategoryIndex, int newCategoryIndex)
 {
-	if (newCategoryIndex<0 || newCategoryIndex>=m_selGroup->categories.size()) {
+	if (newCategoryIndex<0 || newCategoryIndex>=(int)m_selGroup->categories.size()) {
 		return;
 	}
